+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v0.9.3-alpha
+* Removed `EngineState` enumeration:
+    - `ListEnginesAsync(state = "PROVISIONED")`
 ## v0.9.2-alpha
 * Removed `EngineSize` enumeration:
   - `CreateEngineAsync(engineName, size = "XS")`